requirements.txt
Django==4.2.5
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
channels==4.0.0
asgiref==3.6.0
django-debug-toolbar==4.1.0
python-dotenv==1.0.0
coverage==7.6.2
dj-rest-auth==6.0.0
